♻️ Move every core logic of Everest into src/core/everest/*.rs
